refactor: 디렉토리 네임 수정
