+++ src/styles/carousel.scss
